requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.6
matplotlib==3.8.0
pandas==2.1.0